+++ src/connectionMatrix/itemDelegate.hpp
@@ -28,11 +28,14 @@ class ItemDelegate final : public QStyledItemDelegate
 public:
 	ItemDelegate(bool const drawMediaLockedDot, QObject* parent = nullptr);
 	void setDrawMediaLockedDot(bool const drawMediaLockedDot) noexcept;
+	void setDrawCRFAudioConnections(bool const drawCRFAudioConnections) noexcept;
+	bool getDrawCRFAudioConnections() const noexcept;
 
 private:
 	virtual void paint(QPainter* painter, QStyleOptionViewItem const& option, QModelIndex const& index) const override;
 
 	bool _drawMediaLockedDot{ false };
+	bool _drawCRFAudioConnections{ false };
 };
 
 } // namespace connectionMatrix