+++ src/connectionMatrix/view.cpp
@@ -89,6 +89,7 @@ View::View(QWidget* parent)
 	settings->registerSettingObserver(settings::ConnectionMatrix_Transpose.name, this);
 	settings->registerSettingObserver(settings::ConnectionMatrix_ChannelMode.name, this);
 	settings->registerSettingObserver(settings::ConnectionMatrix_ShowMediaLockedDot.name, this);
+	settings->registerSettingObserver(settings::ConnectionMatrix_AllowCRFAudioConnection.name, this);
 	settings->registerSettingObserver(settings::General_ThemeColorIndex.name, this);
 
 	// react on connection completed signals to show error messages.
@@ -105,6 +106,7 @@ View::~View()
 	settings->unregisterSettingObserver(settings::ConnectionMatrix_Transpose.name, this);
 	settings->unregisterSettingObserver(settings::ConnectionMatrix_ChannelMode.name, this);
 	settings->unregisterSettingObserver(settings::ConnectionMatrix_ShowMediaLockedDot.name, this);
+	settings->unregisterSettingObserver(settings::ConnectionMatrix_AllowCRFAudioConnection.name, this);
 	settings->unregisterSettingObserver(settings::General_ThemeColorIndex.name, this);
 }
 
@@ -171,6 +173,12 @@ void View::onIntersectionClicked(QModelIndex const& index)
 				break;
 		}
 	};
+
+	// Do not allow connection for incompatible format types (but allow disconnected)
+	if (!_itemDelegate->getDrawCRFAudioConnections() && intersectionData.flags.test(Model::IntersectionData::Flag::WrongFormatType) && intersectionData.state == Model::IntersectionData::State::NotConnected)
+	{
+		return;
+	}
 
 	switch (intersectionData.type)
 	{
@@ -585,6 +593,13 @@ void View::onSettingChanged(settings::SettingsManager::Setting const& name, QVar
 	{
 		auto const drawDot = value.toBool();
 		_itemDelegate->setDrawMediaLockedDot(drawDot);
+
+		forceFilter();
+	}
+	else if (name == settings::ConnectionMatrix_AllowCRFAudioConnection.name)
+	{
+		auto const drawConnections = value.toBool();
+		_itemDelegate->setDrawCRFAudioConnections(drawConnections);
 
 		forceFilter();
 	}