+++ src/main.cpp
@@ -168,6 +168,7 @@ int main(int argc, char* argv[])
 	settings.registerSetting(settings::ConnectionMatrix_AlwaysShowArrowTip);
 	settings.registerSetting(settings::ConnectionMatrix_AlwaysShowArrowEnd);
 	settings.registerSetting(settings::ConnectionMatrix_ShowMediaLockedDot);
+	settings.registerSetting(settings::ConnectionMatrix_AllowCRFAudioConnection);
 
 	// Network
 	settings.registerSetting(settings::Network_ProtocolType);